arch(policy): slit post rule into basic / stateful category
