+++ api/services/policy/src/engine/rules/AbstractRule.ts
@@ -1,11 +1,12 @@
-import { priority } from '../helpers/priority';
-import { type } from '../helpers/type';
+import { priority, LOWEST } from '../helpers/priority';
+import { type, DEFAULT } from '../helpers/type';
 import { RuleHandlerParamsInterface } from '../interfaces';
 
 export abstract class AbstractRule<P = any> {
+  static readonly type: type = DEFAULT;
+  static readonly priority: priority = LOWEST;
+
   static readonly slug: string;
-  static readonly index: priority;
-  static readonly type: type;
   static readonly schema?: { [k: string]: any };
   static readonly description?: string;
