+++ api/services/policy/src/engine/rules/StatefulRule.ts
@@ -1,10 +1,10 @@
-import { POST } from '../helpers/type';
+import { STATEFUL } from '../helpers/type';
 import { LOWEST } from '../helpers/priority';
 import { RuleHandlerParamsInterface } from '../interfaces';
 import { AbstractRule } from './AbstractRule';
 
-export abstract class PostRule<P = any> extends AbstractRule<P> {
-  static readonly type = POST;
+export abstract class StatefulRule<P = any> extends AbstractRule<P> {
+  static readonly type = STATEFUL;
   static readonly priority = LOWEST;
 
   abstract async apply(context: RuleHandlerParamsInterface): Promise<void>;
