+++ api/services/policy/src/engine/rules/basic/IdfmRegular.ts
@@ -1,5 +1,5 @@
 import { NotApplicableTargetException } from '../../exceptions/NotApplicableTargetException';
-import { PostRule } from '../PostRule';
+import { AbstractRule } from '../AbstractRule';
 import { RuleHandlerParamsInterface } from '../../interfaces';
 
 interface IdfmParametersInterface {
@@ -7,9 +7,9 @@ interface IdfmParametersInterface {
   paris_insee_code: string[];
 }
 
-export class IdfmStrikeJan2020 extends PostRule<IdfmParametersInterface> {
-  static readonly slug: string = 'idfm_strike_jan2020';
-  static readonly description: string = 'Politique de grève IDFM janvier 2020';
+export class IdfmRegular extends AbstractRule<IdfmParametersInterface> {
+  static readonly slug: string = 'idfm_regular';
+  static readonly description: string = 'Politique régulière IDFM';
   static readonly schema: { [k: string]: any } = {
     type: 'object',
     required: ['territory_id', 'paris_insee_code'],
@@ -27,7 +27,7 @@ export class IdfmStrikeJan2020 extends PostRule<IdfmParametersInterface> {
   };
   async apply(ctx: RuleHandlerParamsInterface): Promise<void> {
     if (!ctx.person.is_driver) {
-      throw new NotApplicableTargetException(IdfmStrikeJan2020.slug);
+      throw new NotApplicableTargetException(IdfmRegular.slug);
     }
 
     const eligibleJourneys = ctx.trip.people
@@ -63,10 +63,24 @@ export class IdfmStrikeJan2020 extends PostRule<IdfmParametersInterface> {
       .sort((p1, p2) => (p1.distance > p2.distance ? 1 : p1.distance < p2.distance ? -1 : 0));
 
     if (eligibleJourneys.length === 0) {
-      throw new NotApplicableTargetException(`Campaign "${IdfmStrikeJan2020.slug}" not Application on target`);
+      throw new NotApplicableTargetException(`Campaign "${IdfmRegular.slug}" not Application on target`);
     }
 
-    // give 4€ for any distance
-    ctx.result = 400;
+    let result = 0;
+    // 2km to 15km
+    {
+      // const journeys = eligibleJourneys.filter(p => p.distance < 15000);
+      const distance = eligibleJourneys.reduce((acc, j) => acc + Math.min(15000, j.distance) * j.seats, 0);
+      result += Math.max(150, (10 * distance) / 1000);
+    }
+
+    // 15km to 30km
+    {
+      const journeys = eligibleJourneys.filter((p) => p.distance > 15000);
+      const distance = journeys.reduce((acc, j) => acc + (Math.min(30000, j.distance) - 15000) * j.seats, 0);
+      result += (10 * distance) / 1000;
+    }
+
+    ctx.result = result;
   }
 }
