+++ api/services/policy/src/engine/rules/basic/index.ts
@@ -1,0 +1,5 @@
+import { IdfmRegular } from './IdfmRegular';
+import { IdfmStrikeJan2020 } from './IdfmStrikeJan2020';
+import { StaticRuleInterface } from '../../interfaces';
+
+export const basic: StaticRuleInterface[] = [IdfmRegular, IdfmStrikeJan2020];
