+++ api/services/policy/src/engine/rules/index.ts
@@ -4,8 +4,9 @@ import { StaticRuleInterface } from '../interfaces/RuleInterface';
 import { filters } from './filters';
 import { modifiers } from './modifiers';
 import { setters } from './setters';
-import { posts } from './posts';
+import { stateful } from './stateful';
+import { basic } from './basic';
 import { metas } from './metas';
 import { transformers } from './transformers';
 
-export const rules: StaticRuleInterface[] = [...filters, ...modifiers, ...setters, ...posts, ...metas, ...transformers];
+export const rules: StaticRuleInterface[] = [...filters, ...modifiers, ...setters, ...stateful, ...basic, ...metas, ...transformers];
