+++ api/services/policy/src/engine/rules/stateful/index.ts
@@ -1,0 +1,5 @@
+import { MetaAmountPost } from './MetaAmountPost';
+import { StaticRuleInterface } from '../../interfaces';
+import { MetaTripPost } from './MetaTripPost';
+
+export const stateful: StaticRuleInterface[] = [MetaAmountPost, MetaTripPost];
